requirements.txt
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
opencv-python
torch
torchvision
pandas
openpyxl
Pillow
pygame
tkinter
yolov5